// src/lib/supabase/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      planned_expenses: {
        Row: {
          amount: number
          category_id: string | null
          created_at: string
          currency_code: string
          due_date: string
          id: string
          is_mandatory: boolean
          name: string
          note: string | null
          recurring: boolean
          status: string
          updated_at: string
          user_id: string
        }
        Insert: {
          amount: number
          category_id?: string | null
          created_at?: string
          currency_code?: string
          due_date: string
          id?: string
          is_mandatory?: boolean
          name: string
          note?: string | null
          recurring?: boolean
          status?: string
          updated_at?: string
          user_id: string
        }
        Update: {
          amount?: number
          category_id?: string | null
          created_at?: string
          currency_code?: string
          due_date?: string
          id?: string
          is_mandatory?: boolean
          name?: string
          note?: string | null
          recurring?: boolean
          status?: string
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "planned_expenses_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}